requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.4
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography>=41.0.0
SQLAlchemy==2.0.20